finished docs, renamed namespace
