+++ QuarkClient/app/src/androidTest/java/com/quark/client/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.quarkclient
+package com.quark.client
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
