+++ QuarkClient/app/src/main/java/com/quark/client/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.quarkclient
+package com.quark.client
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.quarkclient.ui.theme.QuarkClientTheme
+import com.quark.client.ui.theme.QuarkClientTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
