+++ QuarkClient/app/src/main/java/com/quark/client/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.quarkclient.ui.theme
+package com.quark.client.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
