+++ doc/proposal.md
@@ -1,10 +1,10 @@
-You should have a proposal document that has a description of your project in some detail. This should be several written paragraphs that expand what you described in the project survey.
+# Quark Proposal
+## Project Description
+Quark is a messaging application that uses quantum-resistant encryption for users to communicate securely and privately. Unlike other messaging apps that use encryption methods which can be broken with quantum computers, Quark uses OQS's quantum-safe encryption through liboqs. Using Quark, your messages will be protected from all potential quantum attacks.
 
-Include in the document a project vision statement (see Chapter 1)
+Using Quark, users send and receive text and image messages, as well as managing their contacts and message history. Additionally, users can customize their app appearance and preferences, such as light mode, dark mode, notifications, etc. You can even choose from a selection of default profile pictures to represent yourself.
 
+Quark is designed for privacy-focused users who value their data and private communication. Quark will always give you the peace of mind that your messages are secure, no matter who you are chatting with. Quark is the ultimate quantum-safe messaging app for the new age of quantum computing. 
 
-
-Quark is a messaging application that uses quantum-resistant encryption. 
-
-Vision statement:
+## Vision Statement
 FOR privacy-focused users WHO need secure communication, Quark is a mobile application THAT utilizes quantum-resistant end-to-end encryption to protect their data. UNLIKE other messaging apps, OUR product offers the highest level of security and privacy in the age of quantum computing.
